+++ dependencies.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 # Dependencies
 
+go get -u github.com/Terry-Mao/gopush-cluster
 go get -u github.com/Terry-Mao/goconf
 go get -u github.com/garyburd/redigo/redis
 go get -u code.google.com/p/go.net/websocket
 go get -u github.com/samuel/go-zookeeper
 go get -u code.google.com/p/log4go
+go get -u github.com/go-sql-driver/mysql
 
 #go get -u ./message/...
 #go get -u ./comet/...